requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
requests==2.31.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.19.0
lxml==4.9.3
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
SQLAlchemy==1.4.36 